requirements.txt
streamlit
pandas
numpy
requests
sentence-transformers
scikit-learn
python-dotenv
openai